CLI-837 move telemetry code to src/core/telemetry
