+++ src/commands/_common/context-augmentation-env.ts
@@ -18,8 +18,9 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
+import { INVOCATION_ID } from '@/core/telemetry/invocation-id.ts';
+
 import { buildSubprocessNetworkEnv } from '../../lib/connectivity/network-config.ts';
-import { INVOCATION_ID } from '../../lib/invocation-id.ts';
 
 export interface ContextAugmentationEnvContext {
   organization?: string;
