+++ src/commands/analyze/sqaa-types.ts
@@ -18,9 +18,10 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
+import type { SqaaTelemetryCallerCommand } from '@/core/telemetry/sqaa-analysis-telemetry.ts';
+
 import type { ResolvedAuth } from '../../lib/auth-resolver.ts';
 import type { SqaaAnalysisDepth } from '../../sonarqube/client.ts';
-import type { SqaaTelemetryCallerCommand } from '../../telemetry/sqaa-analysis-telemetry.ts';
 import type { CloudAuth } from './sqaa-auth.ts';
 import type { SqaaDeepWireDepth } from './sqaa-depth.ts';
