+++ src/commands/hook/agent-post-tool-use.ts
@@ -23,17 +23,18 @@
 
 import { existsSync, readFileSync } from 'node:fs';
 
+import {
+  emitSqaaHookFailureTelemetry,
+  SQAA_CLAUDE_POST_TOOL_USE_CALLER_COMMAND,
+  SQAA_HOOK_TELEMETRY_EXIT_CODE,
+} from '@/core/telemetry/sqaa-analysis-telemetry.ts';
+
 import { resolveAuth } from '../../lib/auth-resolver.ts';
 import { AGENTIC_PACK_URL } from '../../lib/config-constants.ts';
 import { canonicalizePath, toRelativePosixPath } from '../../lib/fs-utils.ts';
 import logger from '../../lib/logger.ts';
 import { timed } from '../../lib/timed.ts';
 import { SqaaForbiddenError } from '../../sonarqube/errors.ts';
-import {
-  emitSqaaHookFailureTelemetry,
-  SQAA_CLAUDE_POST_TOOL_USE_CALLER_COMMAND,
-  SQAA_HOOK_TELEMETRY_EXIT_CODE,
-} from '../../telemetry/sqaa-analysis-telemetry.ts';
 import { resolveSqaaBranch } from '../analyze/sqaa-changeset.ts';
 import { fetchSingleFileReport, finishSqaaTelemetryFromReport } from '../analyze/sqaa-run.ts';
 import { formatSqaaIssuesForHook, writePostToolUseHookOutput } from './format-sqaa-hook-context.ts';
