+++ src/commands/hook/claude-pre-tool-use.ts
@@ -23,8 +23,9 @@
 
 import { existsSync } from 'node:fs';
 
+import { SECRETS_CALLER_COMMANDS } from '@/core/telemetry/secrets-analysis-telemetry.ts';
+
 import logger from '../../lib/logger.ts';
-import { SECRETS_CALLER_COMMANDS } from '../../telemetry/secrets-analysis-telemetry.ts';
 import { EXIT_CODE_SECRETS_FOUND } from '../analyze/secrets.ts';
 import {
   type HookDependencies,
