+++ src/commands/hook/cursor-prompt-submit.ts
@@ -26,8 +26,9 @@
 // `{ "decision": "block", "reason": "..." }`, which is why this handler is separate from
 // `agentPromptSubmit`.
 
+import { SECRETS_CALLER_COMMANDS } from '@/core/telemetry/secrets-analysis-telemetry.ts';
+
 import logger from '../../lib/logger.ts';
-import { SECRETS_CALLER_COMMANDS } from '../../telemetry/secrets-analysis-telemetry.ts';
 import { EXIT_CODE_SECRETS_FOUND } from '../analyze/secrets.ts';
 import {
   type HookDependencies,
