+++ src/commands/hook/git-pre-commit-dependency-risks.ts
@@ -24,15 +24,15 @@
 // missing, scanner failure), and blocks the commit only when risks matching the
 // configured filter are found.
 
+import {
+  emitScaAnalysisTelemetry,
+  SCA_CALLER_COMMANDS,
+} from '@/core/telemetry/sca-analysis-telemetry.ts';
 import { discreetSuccess, success, warn } from '@/core/ui';
 
 import type { ResolvedAuth } from '../../lib/auth-resolver.ts';
 import logger from '../../lib/logger.ts';
 import { SonarQubeClient } from '../../sonarqube/client.ts';
-import {
-  emitScaAnalysisTelemetry,
-  SCA_CALLER_COMMANDS,
-} from '../../telemetry/sca-analysis-telemetry.ts';
 import { CommandFailedError } from '../_common/error.ts';
 import {
   resolveScaScannerBinaryPath,
