+++ src/commands/hook/git-pre-push-secrets.ts
@@ -18,8 +18,9 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
+import { SECRETS_CALLER_COMMANDS } from '@/core/telemetry/secrets-analysis-telemetry.ts';
+
 import type { ResolvedAuth } from '../../lib/auth-resolver.ts';
-import { SECRETS_CALLER_COMMANDS } from '../../telemetry/secrets-analysis-telemetry.ts';
 import { CommandFailedError } from '../_common/error.ts';
 import { resolveSecretsBinaryPath } from '../_common/install/secrets.ts';
 import {
