+++ src/core/telemetry/enabled.ts
@@ -18,19 +18,22 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-import { describe, expect, it } from 'bun:test';
+import { ENV_DO_NOT_TRACK } from '../../lib/config-constants.ts';
+import type { CliState } from '../../lib/state.ts';
 
-import { INVOCATION_ID } from '../../../src/lib/invocation-id';
+/** True when DO_NOT_TRACK is set to 1 */
+export function isDoNotTrackRequested(): boolean {
+  return process.env[ENV_DO_NOT_TRACK]?.trim() === '1';
+}
 
-const UUID_V4_RE = /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
+/** Whether telemetry collection and error reporting should run for this session. */
+export function isTelemetryEnabled(state: CliState): boolean {
+  return state.telemetry.enabled && !isDoNotTrackRequested();
+}
 
-describe('INVOCATION_ID', () => {
-  it('is a v4 UUID string', () => {
-    expect(INVOCATION_ID).toMatch(UUID_V4_RE);
-  });
-
-  it('is stable across re-imports within the same process', async () => {
-    const reimported = (await import('../../../src/lib/invocation-id')).INVOCATION_ID;
-    expect(reimported).toBe(INVOCATION_ID);
-  });
-});
+export function describeTelemetryStatus(state: CliState): string {
+  if (isDoNotTrackRequested()) {
+    return 'Telemetry is currently disabled (DO_NOT_TRACK is set).';
+  }
+  return `Telemetry is currently ${state.telemetry.enabled ? 'enabled' : 'disabled'}.`;
+}
