+++ src/core/telemetry/index.ts
@@ -20,10 +20,10 @@
 
 import { type Command } from 'commander';
 
-import { DISTRIBUTION } from '../lib/distribution.js';
-import { tryLoadState } from '../lib/state-manager.js';
-import { isTelemetryEnabled } from './enabled.js';
-import { emitCommandExecuted, flushTelemetryEvents } from './telemetry-events.js';
+import { DISTRIBUTION } from '../../lib/distribution.ts';
+import { tryLoadState } from '../../lib/state-manager.ts';
+import { isTelemetryEnabled } from './enabled.ts';
+import { emitCommandExecuted, flushTelemetryEvents } from './telemetry-events.ts';
 
 export const TELEMETRY_FLUSH_MODE_ENV = '__SQ_CLI_TELEMETRY_FLUSH__';
