+++ src/core/telemetry/secrets-analysis-telemetry.ts
