+++ tests/unit/commands/hook/codex-post-tool-use.test.ts
@@ -20,15 +20,16 @@
 
 import { afterEach, beforeEach, describe, expect, it, spyOn } from 'bun:test';
 
+import * as sqaaTelemetry from '@/core/telemetry/sqaa-analysis-telemetry.ts';
+import {
+  SQAA_CODEX_POST_TOOL_USE_CALLER_COMMAND,
+  SQAA_HOOK_TELEMETRY_EXIT_CODE,
+} from '@/core/telemetry/sqaa-analysis-telemetry.ts';
+
 import * as sqaaModule from '../../../../src/commands/analyze/sqaa.ts';
 import { codexPostToolUse } from '../../../../src/commands/hook/codex-post-tool-use.ts';
 import * as hookOutput from '../../../../src/commands/hook/format-sqaa-hook-context.ts';
 import * as authResolver from '../../../../src/lib/auth-resolver.ts';
-import * as sqaaTelemetry from '../../../../src/telemetry/sqaa-analysis-telemetry.ts';
-import {
-  SQAA_CODEX_POST_TOOL_USE_CALLER_COMMAND,
-  SQAA_HOOK_TELEMETRY_EXIT_CODE,
-} from '../../../../src/telemetry/sqaa-analysis-telemetry.ts';
 
 describe('codexPostToolUse', () => {
   let stdoutSpy: ReturnType<typeof spyOn>;
