+++ tests/unit/core/telemetry/integrate-telemetry.test.ts
@@ -22,11 +22,11 @@ import { createHash } from 'node:crypto';
 
 import { afterEach, beforeEach, describe, expect, it, spyOn } from 'bun:test';
 
-import type { ResolvedAuth } from '../../../src/lib/auth-resolver.js';
-import { canonicalizePath } from '../../../src/lib/fs-utils.js';
-import type { InstalledIntegrationFeature } from '../../../src/lib/state.js';
-import { emitIntegrationConfiguredTelemetry } from '../../../src/telemetry/integrate-telemetry.js';
-import * as telemetryEvents from '../../../src/telemetry/telemetry-events.js';
+import { emitIntegrationConfiguredTelemetry } from '@/core/telemetry/integrate-telemetry.ts';
+import * as telemetryEvents from '@/core/telemetry/telemetry-events.ts';
+import type { ResolvedAuth } from '@/lib/auth-resolver.ts';
+import { canonicalizePath } from '@/lib/fs-utils.ts';
+import type { InstalledIntegrationFeature } from '@/lib/state.ts';
 
 const AUTH: ResolvedAuth = {
   connectionType: 'cloud',
