+++ tests/unit/core/telemetry/user.test.ts
@@ -25,8 +25,8 @@ import { dirname, join } from 'node:path';
 
 import { afterEach, beforeEach, describe, expect, it, spyOn } from 'bun:test';
 
-import { ENV_SONAR_USER_HOME } from '../../../src/lib/config-constants.js';
-import { getOrCreateUserId } from '../../../src/telemetry/user.js';
+import { getOrCreateUserId } from '@/core/telemetry/user.ts';
+import { ENV_SONAR_USER_HOME } from '@/lib/config-constants.ts';
 
 describe('getOrCreateUserId', () => {
   const previousSonarUserHome = process.env[ENV_SONAR_USER_HOME];
