+++ tests/unit/lib/sentry.test.ts
@@ -24,10 +24,11 @@ import type { ErrorEvent, EventHint } from '@sentry/bun';
 import * as Sentry from '@sentry/bun';
 import { afterEach, beforeEach, describe, expect, it, spyOn } from 'bun:test';
 
+import * as userModule from '@/core/telemetry/user.ts';
+
 import { ENV_DO_NOT_TRACK } from '../../../src/lib/config-constants.js';
 import { flushSentry, initSentry } from '../../../src/lib/sentry.js';
 import { getDefaultState } from '../../../src/lib/state.js';
-import * as userModule from '../../../src/telemetry/user.js';
 
 function makeErrorEvent(
   filenames: (string | undefined)[][],
