+++ tests/unit/lib/update-notification.test.ts
@@ -25,11 +25,11 @@ import { join } from 'node:path';
 import { afterEach, beforeEach, describe, expect, it, spyOn } from 'bun:test';
 import type { Command } from 'commander';
 
+import { COMMAND_TREE } from '@/commands/command-tree.ts';
+import { TELEMETRY_FLUSH_MODE_ENV } from '@/core/telemetry';
 import { setFormattedOutputMode } from '@/core/ui';
 
 import { version as CURRENT_VERSION } from '../../../package.json';
-import { COMMAND_TREE } from '../../../src/commands/command-tree.ts';
-import { TELEMETRY_FLUSH_MODE_ENV } from '../../../src/telemetry/index.js';
 
 const originalEnvForNotify = { ...process.env };
 const tempHome = mkdtempSync(join(tmpdir(), 'sonar-update-notify-'));
